:recycle: Linting and reorganisation of the code
