+++ utils/validation/json-schema-rules.go
@@ -791,109 +791,3 @@ func LengthRule(min, max *int) validation.Rule {
 func RuneLengthRule(min, max *int) validation.Rule {
 	return validation.RuneLength(field.OptionalInt(min, 0), field.OptionalInt(max, math.MaxInt))
 }
-
-func typedSequence[T any](value any) ([]T, error) {
-	v, isNil := validation.Indirect(value)
-	if isNil {
-		return nil, nil
-	}
-	rv := reflect.ValueOf(v)
-	if rv.Kind() != reflect.Array && rv.Kind() != reflect.Slice {
-		return nil, errArrayOrSliceRequired
-	}
-	return collection.MapWithError(collection.Range(0, rv.Len(), nil), func(i int) (T, error) {
-		item, ok := rv.Index(i).Interface().(T)
-		if !ok {
-			var zero T
-			return zero, commonerrors.Newf(commonerrors.ErrMarshalling, "unsupported item type at index [%d]: %T", i, rv.Index(i).Interface())
-		}
-		return item, nil
-	})
-}
-
-func objectValue(value any) (rv reflect.Value, isNil bool, err error) {
-	v, isNil := validation.Indirect(value)
-	if isNil {
-		return reflect.Value{}, true, nil
-	}
-	rv = reflect.ValueOf(v)
-	if rv.Kind() != reflect.Map && rv.Kind() != reflect.Struct {
-		return reflect.Value{}, false, errMapRequired
-	}
-	return rv, false, nil
-}
-
-func hasObjectProperty(rv reflect.Value, key string) bool {
-	switch rv.Kind() {
-	case reflect.Map:
-		return rv.MapIndex(reflect.ValueOf(key)).IsValid()
-	case reflect.Struct:
-		fieldValue := rv.FieldByName(key)
-		return fieldValue.IsValid() && !utilreflection.IsEmpty(fieldValue.Interface())
-	default:
-		return false
-	}
-}
-
-func objectPropertyValue(rv reflect.Value, key string) (any, bool) {
-	switch rv.Kind() {
-	case reflect.Map:
-		value := rv.MapIndex(reflect.ValueOf(key))
-		if !value.IsValid() {
-			return nil, false
-		}
-		return value.Interface(), true
-	case reflect.Struct:
-		fieldValue := rv.FieldByName(key)
-		if !fieldValue.IsValid() {
-			return nil, false
-		}
-		return fieldValue.Interface(), true
-	default:
-		return nil, false
-	}
-}
-
-func objectPropertyNames(rv reflect.Value) []string {
-	switch rv.Kind() {
-	case reflect.Map:
-		keys := rv.MapKeys()
-		return collection.Filter(collection.Map(keys, func(key reflect.Value) string {
-			if key.Kind() != reflect.String {
-				return ""
-			}
-			return key.String()
-		}), func(key string) bool { return key != "" })
-	case reflect.Struct:
-		result := make([]string, 0, rv.NumField())
-		for i := 0; i < rv.NumField(); i++ {
-			result = append(result, rv.Type().Field(i).Name)
-		}
-		return result
-	default:
-		return nil
-	}
-}
-
-func countPresentObjectProperties(rv reflect.Value, keys []string) int {
-	switch rv.Kind() {
-	case reflect.Map:
-		return collection.CountBy(keys, func(key string) bool {
-			mapValue := rv.MapIndex(reflect.ValueOf(key))
-			if !mapValue.IsValid() {
-				return false
-			}
-			return !utilreflection.IsEmpty(mapValue.Interface())
-		})
-	case reflect.Struct:
-		return collection.CountBy(keys, func(key string) bool {
-			fieldValue := rv.FieldByName(key)
-			if !fieldValue.IsValid() {
-				return false
-			}
-			return !utilreflection.IsEmpty(fieldValue.Interface())
-		})
-	default:
-		return 0
-	}
-}
