+++ utils/validation/rules_test.go
@@ -2,9 +2,7 @@ package validation
 
 import (
 	"encoding/base64"
-	"regexp"
 	"testing"
-	"time"
 
 	"github.com/go-faker/faker/v4"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
@@ -126,61 +124,4 @@ func TestIsPathParameter(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestAdditionalGeneralRules(t *testing.T) {
-	t.Run("string rules", func(t *testing.T) {
-		assert.NoError(t, validation.Validate("prefix-value", Prefix("pre")))
-		assert.Error(t, validation.Validate("value", Prefix("pre")))
-
-		assert.NoError(t, validation.Validate("file.txt", Suffix(".txt")))
-		assert.Error(t, validation.Validate("file.log", Suffix(".txt")))
-
-		assert.NoError(t, validation.Validate("hello world", ContainsString("world")))
-		assert.Error(t, validation.Validate("hello", ContainsString("world")))
-
-		re := regexp.MustCompile(`^[a-z]+$`)
-		assert.NoError(t, validation.Validate("hello", Like(re)))
-		assert.Error(t, validation.Validate("123", Like(re)))
-
-		assert.NoError(t, validation.Validate("hello", NotContains("world")))
-		assert.Error(t, validation.Validate("hello world", NotContains("world")))
-
-		assert.NoError(t, validation.Validate("helloworld", NotContainsWhitespaces()))
-		assert.Error(t, validation.Validate("hello world", NotContainsWhitespaces()))
-	})
-
-	t.Run("length and occurs", func(t *testing.T) {
-		assert.NoError(t, validation.Validate("ab", LengthExact(2)))
-		assert.Error(t, validation.Validate("abc", LengthExact(2)))
-
-		assert.NoError(t, validation.Validate("banana", MinOccurs("a", 3)))
-		assert.Error(t, validation.Validate("banana", MinOccurs("a", 4)))
-
-		assert.NoError(t, validation.Validate("banana", MaxOccurs("a", 3)))
-		assert.Error(t, validation.Validate("banana", MaxOccurs("a", 2)))
-
-		assert.NoError(t, validation.Validate("banana", OccursExactly("a", 3)))
-		assert.Error(t, validation.Validate("banana", OccursExactly("a", 2)))
-	})
-
-	t.Run("time rules", func(t *testing.T) {
-		assert.NoError(t, validation.Validate("5s", IsDuration))
-		assert.Error(t, validation.Validate("not-a-duration", IsDuration))
-		assert.NoError(t, validation.Validate("10s", DurationMinimum(5*time.Second)))
-		assert.Error(t, validation.Validate("1s", DurationMinimum(5*time.Second)))
-		assert.NoError(t, validation.Validate("5s", DurationConst(5*time.Second)))
-		assert.Error(t, validation.Validate("6s", DurationConst(5*time.Second)))
-
-		assert.NoError(t, validation.Validate("2024-01-01T00:00:00Z", IsTimestamp))
-		assert.Error(t, validation.Validate("2024-01-01", IsTimestamp))
-		minTime := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		maxTime := time.Date(2024, 12, 31, 0, 0, 0, 0, time.UTC)
-		assert.NoError(t, validation.Validate("2024-06-01T00:00:00Z", TimestampMinimum(minTime)))
-		assert.Error(t, validation.Validate("2023-12-31T00:00:00Z", TimestampMinimum(minTime)))
-		assert.NoError(t, validation.Validate("2024-06-01T00:00:00Z", TimestampMaximum(maxTime)))
-		assert.Error(t, validation.Validate("2025-01-01T00:00:00Z", TimestampMaximum(maxTime)))
-		assert.NoError(t, validation.Validate("2024-01-01T00:00:00Z", TimestampConst(minTime)))
-		assert.Error(t, validation.Validate("2024-01-02T00:00:00Z", TimestampConst(minTime)))
-	})
 }
